README.md
An important detail regarding the status of the engineer visa and the international services/specialist in humanities has been added to reflect the legal amendments made in 2015 April by the Japanese governement. The 2 aforementioned visa's have been combined to form 1 visa called the 'Engineer/Specialist in Humanites/International Services Visa.